added sanctum and article policy

+++ app/Repositories/ArticleRepository.php
@@ -5,6 +5,7 @@ namespace App\Repositories;
 use App\Models\Article;
 use Illuminate\Http\Request;
 use Illuminate\Pagination\LengthAwarePaginator;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleRepository implements ArticleRepositoryInterface
 {
@@ -51,7 +52,43 @@ class ArticleRepository implements ArticleRepositoryInterface
      */
     public function create(array $data)
     {
-        return Article::create($data);
+        if (Gate::denies('create-article')) {
+            abort(403, 'Unauthorized');
+        }
+
+        // Create a new article for 'uk' locale
+        $articleUk = new Article('uk');
+        $articleUk->title = $data['title'];
+        $articleUk->slug = $data['slug'];
+        $articleUk->date = $data['date'];
+        $articleUk->excerpt = $data['excerpt'] ?? '';
+        $articleUk->image = $data['image'] ?? '';
+        $articleUk->id_article_tags = $data['id_article_tags'];
+        $articleUk->id_article_categories = $data['id_article_categories'];
+        $articleUk->content = $data['content'];
+        $articleUk->is_show = $data['is_show'];
+        $articleUk->meta_title = $data['meta_title'] ?? '';
+        $articleUk->meta_description = $data['meta_description'] ?? '';
+        $articleUk->meta_keywords = $data['meta_keywords'] ?? '';
+        $articleUk->save();
+        
+        // Create a new article for 'en' locale
+        $articleEn = new Article('en');
+        $articleEn->title = $data['title'];
+        $articleEn->slug = $data['slug'];
+        $articleEn->date = $data['date'];
+        $articleEn->excerpt = $data['excerpt'] ?? '';
+        $articleEn->image = $data['image'] ?? '';
+        $articleEn->id_article_tags = $data['id_article_tags'];
+        $articleEn->id_article_categories = $data['id_article_categories'];
+        $articleEn->content = $data['content'];
+        $articleEn->is_show = $data['is_show'];
+        $articleEn->meta_title = $data['meta_title'] ?? '';
+        $articleEn->meta_description = $data['meta_description'] ?? '';
+        $articleEn->meta_keywords = $data['meta_keywords'] ?? '';
+        $articleEn->save();
+        
+        return $articleUk;
     }
 
     /**
@@ -59,6 +96,10 @@ class ArticleRepository implements ArticleRepositoryInterface
      */
     public function update(int $id, array $data)
     {
+        if (Gate::denies('update-article')) {
+            abort(403, 'Unauthorized');
+        }
+
         $article = Article::findOrFail($id);
         $article->update($data);
 
@@ -70,6 +111,10 @@ class ArticleRepository implements ArticleRepositoryInterface
      */
     public function delete(int $id)
     {
+        if (Gate::denies('delete-article')) {
+            abort(403, 'Unauthorized');
+        }
+
         $article = Article::findOrFail($id);
         return $article->delete();
     }
